dev: make free_pages() get void *
and update callsites with spatch:

@@
identifier addr;
expression order;
@@
(
- free_pages(addr, order)
+ free_pages((void *)addr, order)
|
- free_page(addr)
+ free_page((void *)addr)
)

@@
type T != void *;
expression addr, order;
@@
(
- free_pages((T)addr, order)
+ free_pages(addr, order)
|
- free_page((T)addr)
+ free_page(addr)
)

+++ arch/powerpc/platforms/cell/spufs/inode.c
@@ -657,8 +657,7 @@ static int spufs_parse_param(struct fs_context *fc, struct fs_parameter *param)
 
 static void spufs_exit_isolated_loader(void)
 {
-	free_pages((unsigned long) isolated_loader,
-			get_order(isolated_loader_size));
+	free_pages(isolated_loader, get_order(isolated_loader_size));
 }
 
 static void __init
