+++ drivers/gpu/drm/panthor/panthor_mmu.c
@@ -491,7 +491,7 @@ static void free_pt(void *cookie, void *data, size_t size)
 	struct panthor_vm *vm = cookie;
 
 	if (unlikely(vm->root_page_table == data)) {
-		free_pages((unsigned long)data, get_order(size));
+		free_pages(data, get_order(size));
 		vm->root_page_table = NULL;
 		return;
 	}
