+++ drivers/misc/ocxl/link.c
@@ -376,7 +376,7 @@ static void free_spa(struct ocxl_link *link)
 		link->dev);
 
 	if (spa && spa->spa_mem) {
-		free_pages((unsigned long) spa->spa_mem, spa->spa_order);
+		free_pages(spa->spa_mem, spa->spa_order);
 		kfree(spa);
 		link->spa = NULL;
 	}
