+++ mm/kasan/shadow.c
@@ -390,7 +390,7 @@ static int __kasan_populate_vmalloc(unsigned long start, unsigned long end, gfp_
 		nr_total -= nr_pages;
 	}
 
-	free_page((unsigned long)data.pages);
+	free_page(data.pages);
 
 	return ret;
 }
